src/test/java/de/tum/in/ase/eist/QueryProcessorTest.java
package de.tum.in.ase.eist;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.fail;

class QueryProcessorTest {

	private final QueryProcessor queryProcessor = new QueryProcessor();

	@Test
	void testEmptyStringIfCannotProcessQuery() {
		assertEquals("", queryProcessor.process("test"));
	}

	@Test
	void testKnowsAboutShakespeare() {
		String actual = queryProcessor.process("Shakespeare");
		if (!actual.contains("playwright")) {
			fail("The QueryProcessor does not know about Shakespeare.");
		}
	}

	@Test
	void isNotCaseSensitive() {
		String actual = queryProcessor.process("shakespeare");
		if (!actual.contains("playwright")) {
			fail("Your QueryProcessor should not be case sensitive.");
		}
	}

	@Test
	void testSummation() {
		String actual = queryProcessor.process("xy: what is 12 plus 3");
		assertEquals("15", actual);
	}

	// which of the following numbers is the largest: 208, 95
	@Test void testLargest() {
		String actual = queryProcessor.process("xy: which of the following numbers is the largest: 208, 95");
		assertEquals("208", actual);
	}

}
